ref: move namespaces around
