+++ src/Server/Game/CharacterObjectLoader.cs
@@ -6,8 +6,8 @@
 using System;
 using System.Collections.Generic;
 using Imlight.Common.Utilities;
-using Imlight.Server.Data;
 using Imlight.Server.Game.Models;
+using Imlight.Server.Shared.Resources;
 using WizUnraveler.ObjectProperty;
 using static WizUnraveler.Cache.TypeCache;
