+++ src/Server/Game/GameWorld.cs
@@ -6,9 +6,9 @@
 using System.Collections.Generic;
 using Akka.Actor;
 using Imlight.Common.Utilities;
-using Imlight.Server.Data;
 using Imlight.Server.Shared.Networking;
 using Imlight.Server.Shared.Packets;
+using Imlight.Server.Shared.Resources;
 
 namespace Imlight.Server.Game
 {
