+++ src/Server/Game/Zone/WizardZoneLoader.cs
@@ -9,9 +9,10 @@ using System.Linq;
 using System.Text.RegularExpressions;
 using Akka.Actor;
 using Imlight.Common.Utilities;
-using Imlight.Server.Data;
-using Imlight.Server.Data.WizardData;
 using Imlight.Server.Shared.Packets;
+using Imlight.Server.Shared.Resources;
+using Imlight.Server.WizardData;
+using Imlight.Server.WizardData.Collections;
 using SharpDX;
 using WizUnraveler.Cache;
 using WizUnraveler.Formats;
