+++ src/Server/Game/Zone/WizardZonePath.cs
@@ -9,9 +9,9 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Akka.Actor;
-using Imlight.Server.Data;
 using Imlight.Server.Shared.Networking;
 using Imlight.Server.Shared.Packets;
+using Imlight.Server.Shared.Resources;
 using WizUnraveler.IO;
 using WizUnraveler.ObjectProperty;
 using static WizUnraveler.Cache.TypeCache;
