+++ src/Server/Imlight.Server.csproj
@@ -9,6 +9,7 @@
         <PackageReference Include="Akka.Remote" Version="1.5.0" />
         <PackageReference Include="Faker.Net" Version="2.0.154" />
         <PackageReference Include="LiteDB" Version="5.0.16" />
+        <PackageReference Include="RavenDB.Client" Version="5.4.109" />
     </ItemGroup>
 
     <ItemGroup>
