+++ src/Server/Login/Services/AuthenticatorService.cs
@@ -12,9 +12,9 @@ using Akka.Actor;
 using WizUnraveler.DML;
 using WizUnraveler.Cache;
 using Imlight.Common.Utilities;
-using Imlight.Server.Data;
 using Imlight.Server.Shared.Networking;
 using Imlight.Server.Shared.Packets;
+using Imlight.Server.Shared.Resources;
 
 namespace Imlight.Server.Login.Services
 {
