+++ src/Server/WizardData/WizardZoneData.cs
@@ -5,7 +5,7 @@
 
 using System.Collections.Generic;
 
-namespace Imlight.Server.Data.WizardData;
+namespace Imlight.Server.WizardData;
 
 public class WizardZoneData
 {
